+++ atacseq/ccbr_trim_align/ccbr_bam_filter_by_mapq.py
@@ -36,5 +36,3 @@ for read in samfile.fetch():
 		pairedreads.write(read)
 samfile.close()
 pairedreads.close()
-
-