+++ atacseq/ccbr_trim_align/ccbr_cutadapt_pe.bash
@@ -1,11 +1,11 @@
 #!/bin/bash
 
-. /opt/conda/etc/profile.d/conda.sh
+. /opt2/conda/etc/profile.d/conda.sh
 conda activate python3
 
 set -e -x -o pipefail
 ARGPARSE_DESCRIPTION="Trim PE reads using cutadapt"      # this is optional
-source /opt/argparse.bash || exit 1
+source /opt2/argparse.bash || exit 1
 argparse "$@" <<EOF || exit 1
 parser.add_argument('--infastq1',required=True, help='input R1 fastq.gz file')
 parser.add_argument('--infastq2',required=True, help='input R2 fastq.gz file')
@@ -29,8 +29,8 @@ cutadapt \
 -n 5 -O 5 \
 -q 10,10 \
 -m 35:35 \
--b file:/opt/TruSeq_and_nextera_adapters.consolidated.fa \
--B file:/opt/TruSeq_and_nextera_adapters.consolidated.fa \
+-b file:/opt2/TruSeq_and_nextera_adapters.consolidated.fa \
+-B file:/opt2/TruSeq_and_nextera_adapters.consolidated.fa \
 -j $ncpus \
 -o $outfastq1 \
 -p $outfastq2 \