+++ atacseq/ccbr_trim_align/install_R_packages.R
@@ -1,0 +1,4 @@
+install.packages("plotly", repos="http://cran.us.r-project.org")
+install.packages("RColorBrewer", repos="http://cran.us.r-project.org")
+install.packages("reshape2", repos="http://cran.us.r-project.org")
+install.packages("rmarkdown", repos="http://cran.us.r-project.org")